+++ crates/lkjscript-core/src/bytecode/structural/references.rs
@@ -1,14 +1,14 @@
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct StructuralDestinationFieldRef {
     pub destination: StructuralDestinationId,
-    pub field: u16,
+    pub field: u64,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct StructuralAggregateFieldRef {
     pub representation: StructuralRepresentationId,
     pub active_variant: Option<VariantId>,
-    pub field: u16,
+    pub field: u64,
     pub result: StructuralFieldMetadata,
     pub result_representation: Option<StructuralRepresentationId>,
 }
